drop tests that seem to test upstreamed functionality
